requirements.txt
pandas
numpy
matplotlib
seaborn
nltk
textblob
scikit-learn
python-dotenv
vaderSentiment